requirements.txt
affinegap==1.11
attrs==19.3.0
backcall==0.1.0
backports.csv==1.0.7
bleach==3.1.0
blis==0.4.1
BTrees==4.6.1
cachetools==4.0.0
catalogue==1.0.0
categorical-distance==1.9
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
csvdedupe==0.1.18
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
datetime-distance==0.1.3
de-core-news-sm==2.2.5
decorator==4.4.1
dedupe==1.10.0
dedupe-hcluster==0.3.8
dedupe-variable-datetime==0.1.5
defusedxml==0.6.0
DoubleMetaphone==0.1
en-core-web-sm==2.2.5
entrypoints==0.3
fastcluster==1.1.26
funcy==1.14
future==0.18.2
haversine==2.2.0
highered==0.2.1
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
jellyfish==0.7.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
kiwisolver==1.1.0
langdetect==1.0.7
Levenshtein-search==1.4.5
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
persistent==4.5.1
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
pyemd==0.5.1
Pygments==2.5.2
pyhacrf-datamade==0.2.5
PyLBFGS==0.2.0.13
pyLDAvis==2.1.2
pyparsing==2.4.6
Pyphen==0.9.5
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rlr==2.4.5
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
simplecosine==1.2
six==1.14.0
spacy==2.2.3
spacy-langdetect==0.1.2
srsly==1.0.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
textacy==0.9.1
thinc==7.3.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==2.2.0
zope.index==5.0.0
zope.interface==4.7.1
